00-mcp/pyproject.toml
[project]
name = "00-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-ai-inference>=1.0.0b9",
    "dotenv>=0.9.9",
    "mcp[cli]>=1.26.0",
]
